requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
langchain>=0.2.12
langchain-core>=0.2.33
langchain-openai>=0.1.21
langgraph>=0.2.3
faiss-cpu>=1.8.0
sentence-transformers>=3.0.1
yt-dlp>=2024.8.6
webvtt-py>=0.4.6
pypdf2>=3.0.1
docx2txt>=0.8
